Envoyer le pwd  non chiffré à PASTELL
